requirements.txt
blinker==1.6.2
cachelib==0.10.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
diffusers==0.16.1
filelock==3.12.0
Flask==2.3.2
Flask-Session==0.5.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
sympy==1.12
tokenizers==0.13.3
torch==2.1.0.dev20230531
torchaudio==2.1.0.dev20230531
torchvision==0.16.0.dev20230531
tqdm==4.65.0
transformers
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
